Created first version of 'Read and Write a netCDF File' notebook.
This notebook demonstrates how to read a netCDF file into a Xarray dataset and how to view a summary of the dataset. A brief introduction to Xarray data arrays and datasets is included. The notebook concludes by writing the dataset to a different netCDF file.

Created and added the ds_BM_NP.nc dataset.

This dataset was created using the DEA's Sandbox and covers the Grose Valley in the Blue Mountains National Park for the period 2018 - 2020; Sentinel 2AB resampled to months using the median.

A hyperlink from the 'Introduction' notebook to the 'Read and Write a netCDF File' notebook was created.